HydrAPI/requirements.txt
requests
supabase
dotenv
quart
quart-cors
python-dotenv
numpy
pytz
PyJWT
bcrypt